Clean up the Bridge...
It is often the case that when a choice is made to perform blocking or non-blocking activities in an Actor, it ends up being better to allow it to block.

What happens, in a non-blocking situation, when another request comes in for the same item while we are awaiting completion of the action in a parallel worker. It is often incorrect to allow two successive requests to the blocking activity. So just better to allow the Actor to block on the first one.