code correction
No matter which thread actually effects the RMW modification, all threads begin with the same tree, and hence compute the same answer stored in the ANS binding.

We should not constantly reset ANS on entry to the mutator function. That could interfere with original caller thread which expect to see a definite ANS on its return from RWM.