Update actors.lisp
In order to get the greatest speed from the RUN dispatcher, everything in it has been honed to a 'T'. It blindly assumes that all arguments are valid, e.g., that the first item in an event message is an Actor.

When this isn't valid, the Lisp system generates nasty machine fault errors. No harm done, but just the same...

So we take precautions in SEND, BECOME, and elsewhere to avoid ever placing invalid event messages into the RUN event queue.