overhaul for better SMP behavior
Using a prefix PAR-SAFE is non-optimal, and frequently not what you want.

We removed SEND, BECOME as generally available functions. They should not be. They should only be alive inside a running Actor code.

Furthermore, when BECOME is used,  it needs to decide which sponsor makes the most sense to run on.

SEND and USING-BECOME are available to code inside an ACTOR form, a DEF-BEH form, or a BEHAVIOR form.

BECOME is only available inside a USING-BECOME form. USING-BECOME specifies which sponsor should be used, and arranges for retransmission of the message in that sponsor when it is not already there.