let's back up here...
Getting another case of 2nd system syndrome.

Instead of playing constantly defensive, how about we expect the foreign sponsor code to be wrapped with a thingy that returns results to us back in our original sponsor.

If you don't trust that the actors have been declared with (IOREQ (IO <actor>)), then wrap it yourself before sending to it.

IOREQ is idempotent.