finish conversion to repudiable signatures
signatures and messages are fully repudiable since we publish the prior signature keying along with the new encryption and its signature.

Our encryption is malleable. So anyone could have later intercepted the message and changed both the encrypted content and signature.

Hence it cannot be proven that either of us in the conversation sent the message.