protect against attacks...
With malleable encryption and repudiable signatures, it becomes possible for an attack to mount a replay attack using a mutated encrypted message and valid signature.

But mutating the cryptotext is likely to produce a gibberish marshal encoding, and if it doesn't, then surely it will produce a different checksum hash on the decoded result.