new round of experiments - fixups for SDLE Store
What happens if we make Actors a funcallable standard object? We could unify the syntax of SEND and FUNCALL. And by making global Actors as first-class functions we also have first-position rights to sending messages...

So SEND and FUNCALL are only necessary when active object is not in first position, or when they have been lexically rebound.

Is this a good or a bad idea? Let's find out...