clean up our act...
No direct mutation of Actor beh slots, except by Dispatcher.

If Actors are provided to ACTORS macro, instead of behavior functional closures, the new Actors become FWD Actors - not stealing a copy of their behaviors.

This is all more in keeping with the immutable FPL style, and not having inadvertent copies of things that ought to exist as single instances.

Slightly more cost for the safety of FPL. Maybe you will thank me someday...