Massive simplification
Invent Special Actors, which are rare and generally not overtly constructed by users.

A Special Actor is one which serves as a customer who absolutely needs to have a response.

Examples are the interposer customer for Serializers and UNW-PROT.

Move the error trapping into the dispatcher, for essentially zero overhead cost, and relax the requirements on Actors which serve as targets of Serializers and UNW-PROT.

Missing match clauses, which cause the message to be dropped on the floor, now have the same character as flaky network connections. Messages just disappear sometimes and are never delivered.

Same too, for customers which get stashed into waiting queues and are forgotten.

Those cases require timers to send timeout exceptions to them. Welcome to the real world...

Errors are caught by the dispatcher, and if a message is seen to have its first argument as a Special Actor, that Actor will be sent the exception condition object - which serves as a response.

Non-special customer Actors are simply forgotten in the event of an ERROR. It becomes as though the message were never delivered.

If you absolutely need a response, then use a SEND-AFTER to send your customer a timeout, or else construct a Special Actor to forward a message to the original customer, and be sure to make it into a ONCE-like behavior, (BECOME-SINK) after handling a message, and also set up a SEND-AFTER to the Special Actor on timeout.

Normal responses from target Actors serve the purpose as always.

A Special Actor should be a ONCE forwarding Actor interposing between target Actors and the original customer Actor. See the example for SPECIAL-TAG in Serializer.lisp.

This change allows a significant simplification to the code base, and so some macros have been discarded. DEF-BEH -- just use DEFUN with ALAMBDA
WITH-ERROR-RESPONSE -- no longer needed
DEF-SER-BEHAVIOR -- no longer needed

Feel free to implement your own catchall clauses for message, instead of just dropping them on the floor. That way you won't have to wait for the timeout to happen.

When, and if, you ever produce a Special Actor, be sure to implement a SEND-AFTER alongside, to serve as a timeout timer.