Update par-xlate.lisp
There is a distinct difference between using LAMBDA and ALAMBDA.

With LAMBDA , if your arguments don't match the declared lambda arglist you will generate a runtime error.

With ALAMBDA, anything goes, and if you aren't matched by a clause it just drops the message on the floor with no notification.