Try something new...
By making Actors into CLOS Funcallable Instances we get automatic forwarding whenever a behavior is set to another Actor.

But, alas, there is a minor cost. CAS needs something like a CONS cell, not a CLOS Instance Slot.

FWIW, Actors and Functions are on a more even keel. Sending to a function causes the function to be executed at commit time. Likewise, funcalling an Actor causes it to stage a Send.