Elide the :CANARY from the AES packet.
The :CANARY provides a cross check on the correctness of the AES decryption. But it also serves as a hint to any attackers that they have succeeded.

If there are potentially several decryption keys to be tried by an attacker, this tells them they found the solution and can quit looking.

By eliding the :CANARY, all other elements of the packet are just random values and provide no hint of correctness.

In our network protocol, if any error occurs during the initial handshake dance, then the nodes go quiet and shutdown.

You get only one chance to answer correctly.