Invent non-destructive encrypt/decrypt
We wanted to avoid needless copying of large buffers. And most of the time this is fine because the encryptors are used in a pipeline where the buffer was freshly invented.

But now comes kvdb with encrypted storage and the buffer being decrypted belongs to kvdb, not freshly allocated.

FPL hygiene, or lack thereof, finally bit us.