Double-Ratchet done correctly
The previous Advancing-Encr/Decr could only handle strict ABAB conversatoins. It fails under a flood of transmissions that happen before a response can be seen, for advancing the ratchet encryption.

As a result, a flood of transmissions caused the system to eventually fall out of sync and lock the channel. It would eventually close due to inactivity across the network.

This new version handles floods and non-strict ABAB conversations just fine. Messages with stale encryption are allowed, and eventually the flood will subside and the ratcheted crypto will take hold for new transmissions.