Fix corner case...
Found a corner case where the system is freshly booted - hence no Actors running yet.

An ASK is performed, which targets a FORK of two services, both of which may block waiting on a mailbox for private comms with CAPI.

Since there are no other threads to take over during the mailbox wait, the system fails to complete the FORK.

Luckily, I had a SERIALIZER with a timeout waiting to fire. That timeout fired, completing the ASK, and so the system did not hang. But it failed to complete the CAPI graphics init of one of the two panes involved.

So as a precaution, if an ASK is performed, we ensure that the Actors system is up and running before becoming one of the dispatchers.