I finally found it...
This code now performs an absolute minimum number of modulo operations.

It happens when you ask for the INT value of a FFLD instance, or when you explicitly normalize with FF-NORMALIZE.

Otherwise, it allows for the indefinite buildup of modular excesses during Add / Subtract operations. During Multiply, it is simply necessary to trim back the excess to no more than the containing binary size of the modular field.

That is, if 2^(n-1) < q < 2^n, for prime modulus q, we simply ensure that products get wrapped to no more than 2^n bits.

Otherwise, during large exponentiations, which happens for Sqrt operations, the buildup of excess, even if only a single bit every now and then, becomes impossibly large when exponents are on the order of 2^250 bits.