backtrack on Serializer timeouts
Serializers are expressly intended for code that cannot fail to respond to the Serializer customer.

Timeouts belong to the customer, if they want them. Serializers should never be used on systems that may fail to produce a response to the customer. This failure could be in the eye of the beholder because it is impossible to discern whether a failure occurs becuase the code takes too long to perform, or else a subsystem has gone off-line.

If a Serializer Actor chain fails, then the entire system should also fail with it - as in power outage.

If you need more than this, then it becomes a less tractible problem needing more power than a Serializer can offer.

If an Actor chain behind a Serializer fails to respond to the customer then that is a programming error and needs to be tracked down and corrected.